pyproject.toml
[project]
name = "robotframework-concurrent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "interpreters-pep-734>=0.5.0",
    "robotframework>=7.1.1",
]

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "radon>=6.0.1",
    "ruff>=0.8.4",
]

interpreters = ["interpreters-pep-734"]

[tool.hatch.version]
source = "versioningit"
default-version = "0.0.0+unknown"

[tool.hatch.version.format]
distance = "{next_version}.dev{distance}+{vcs}{rev}"
dirty = "{version}+dirty"
distance-dirty = "{next_version}.dev{distance}+{vcs}{rev}.dirty"

[build-system]
requires = ["hatchling", "versioningit"]
build-backend = "hatchling.build"
